Fix for writing large XML requests
Net::SSH2::Channel2::write() isn't guaranteed to write the entire buffer
that is passed to it, especially if the buffer is large. Therefore loop
over it until all bytes have been written.

In order to avoid busy-looping while the channel isn't ready to accept
more data, set it as blocking before starting the write loop. Make
recv() explicitly set the channel back to non-blocking before starting
the read loop. This removes the need to set the blocking state in
start().