Improved handling of disconnect / multiple instances and non-listed methods (#34)
* adding exception handling for reads and writes

I've added exception handling to catch when a write or read
fails. When one of these failures occur, we disconnect the underlying
ssh connection. It is the callers responsibility to attempt any
reconnections.

* adding ref to methods on object for external access

* Removing static error messages

Error messages were persisting as new Device objects were
created. This caused unexpected errors to be announced.

* adding spec file to build rpm

* adding two extra dependencies

* updating version and spec file with a fix for our dying NetConf session....

* updated the build process and cleaned up a few things for Juniper folks

* One last fix for when we receive 0 bytes because the netconf session has died

* fixing the timeout so if we do get data and it takes more than 15sec to receive it we wont timeout